% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarities.R
\name{asim_centroid}
\alias{asim_centroid}
\alias{asim_semantic_min_sum}
\alias{asim_semantic_min_max}
\alias{asim_rwmd}
\title{Aggregation similarities for sparse 3d representations}
\usage{
asim_centroid(x, y = NULL, vecs, ptrans = NULL, strans = NULL,
  by = c("primary", "secondary", "row", "column"), precompute = TRUE,
  dist_type = "cosine")

asim_semantic_min_sum(x, y = NULL, vecs, ptrans = NULL, strans = NULL,
  by = c("primary", "secondary", "row", "column"), precompute = TRUE,
  dist_type = "cosine")

asim_semantic_min_max(x, y = NULL, vecs, ptrans = NULL, strans = NULL,
  by = c("primary", "secondary", "row", "column"), precompute = TRUE,
  dist_type = "cosine")

asim_rwmd(x, y = NULL, vecs, ptrans = NULL, strans = NULL,
  by = c("primary", "secondary", "row", "column"), precompute = TRUE,
  dist_type = "cosine")
}
\arguments{
\item{x}{sparse or dense objects supported by \code{simdist}. See
\code{\link{simdist-package}} for description of supported types and
related terminology.}

\item{y}{sparse or dense objects supported by \code{simdist}. See
\code{\link{simdist-package}} for description of supported types and
related terminology.}

\item{vecs}{Dense matrix with columns}

\item{ptrans}{Primary and secondary transformations. Can be either a
function, string or a numeric vector. When a function, it must take 3
arguments - an object supported by \code{simdist} distance measures,
names of a dimension ("primary" or "secondary") and logical na.rm. When
a string it must name a transformation function without the semantic
prefix (norm_, scale_ or trans_); see \code{link{transformations}}. When
numeric vector it specifies weights to scale along the corresponding
dimension.}

\item{strans}{Primary and secondary transformations. Can be either a
function, string or a numeric vector. When a function, it must take 3
arguments - an object supported by \code{simdist} distance measures,
names of a dimension ("primary" or "secondary") and logical na.rm. When
a string it must name a transformation function without the semantic
prefix (norm_, scale_ or trans_); see \code{link{transformations}}. When
numeric vector it specifies weights to scale along the corresponding
dimension.}

\item{by}{Dimension along which to perform distance computation. For all
supported data structures computation along primary dimension is more or
as efficient than along the secondary dimension.}

\item{precompute}{logical Weather to optimize the computation for speed and
precompute individual distances. The computation is method specific bug
generally should be \code{TRUE} (the default) unless memory usage is a
concern.}

\item{dist_type}{distance to use across individual vectors in \code{vecs}}
}
\description{
These are similarity counterparts of \code{aggregated_distances}, which
see. Generally similarities are computed from their distance counterparts
through \code{s=1-d} or \code{s=1/(1+d)}, but this need not always be the
case. See the source code for the exact computation.
}
