% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{dist_cosine}
\alias{dist_cosine}
\alias{dist_euclidean}
\alias{distances}
\title{Distances for sparse 2d representations}
\usage{
dist_cosine(x, y = NULL, ptrans = "l2", strans = NULL, by = c("primary",
  "secondary", "row", "column"), ref_names = NULL, pairwise = FALSE)

dist_euclidean(x, y = NULL, ptrans = NULL, strans = NULL,
  by = c("primary", "secondary", "row", "column"), ref_names = NULL,
  pairwise = FALSE)
}
\arguments{
\item{x, y}{sparse or dense objects supported by \code{simdist}. See
\code{\link{simdist-package}} for description of supported types and
related terminology.}

\item{ptrans, strans}{Primary and secondary transformations. Can be either a
function, string or a numeric vector. When a function, it must take 3
arguments - an object supported by \code{simdist} distance measures,
names of a dimension ("primary" or "secondary") and logical na.rm. When
a string it must name a transformation function without the semantic
prefix (norm_, scale_ or trans_); see \code{link{transformations}}. When
numeric vector it specifies weights to scale along the corresponding
dimension.}

\item{by}{Dimension along which to perform distance computation. For all
supported data structures computation along primary dimension is more or
as efficient than along the secondary dimension.}

\item{ref_names}{reference names for secondary entries. By default both
\code{x} and \code{y} secondary names are used. Secondary entries not in
\code{ref_names} are discarded from \code{x} and \code{y}.}
}
\description{
Distances for sparse 2d representations
}

