% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{norm_l1}
\alias{norm_l2}
\alias{transform}
\title{Transformations across primary and secondary dimensions}
\usage{
norm_l1(x, by = c("primary", "secondary", "row", "column"), na.rm = FALSE)

norm_l2(x, by = c("primary", "secondary", "row", "column"), na.rm = FALSE)
}
\arguments{
\item{x}{object}

\item{by}{name of the dimension}

\item{na.rm}{logical}
}
\description{
Transformations across primary and secondary dimensions
}

