% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat2psv}
\alias{mat2psv}
\alias{psv}
\alias{psv2mat}
\title{Various utilities}
\usage{
mat2psv(mat)

psv2mat(df)

psv(df, pname = NULL, sname = NULL, vname = NULL)
}
\arguments{
\item{mat}{matrix}

\item{df}{data.frame}

\item{pname}{primary id name}

\item{sname}{secondary id name}

\item{vname}{value name}
}
\description{
\code{psv}: mark a data frame with attributes in order to track which are
primary, seocndary and value columns.
}

